Added overflow handler
The overflow handler handles, well, value overflows that may happen when over-using certain action. For example, the player's hydration can't go beyond 100%, thus it can utilize the CAP overflow handler. On the other hand, heat can easily exceed 100% of the maximum value, and so we can use the NONE overflow handler. There is also a so-called WARP handler, which is currently unused. It may come handy one day.