Fixed wetness slowdown in MP
Due to PlayerAttributeModifer and optimization attempts, the wetness was bugged in MP so that it applied the slowness effect of the last player on which the attribute was processed. This was fixed by a rather ugly hack. But maybe it was intended that way. I dunno. But anyways I think it causes way too much List I/O operations every tick.