Reworked the towel item
The towel item has been rewritten from the ground up. Now, the towel takes semi-realistic approach to using. It tries to equalize it's wetness with that of the player so that when you use wet towel on yourself it doesn't absorb any more wetness that it is possible. In other words, the towel, when used, absorbs water from the player so that their relative wetness scale is roughly equivalent.