Changed stat update handling
This is a HUGE update. Until now, stats were updated by packets sent by the server EACH tick. This not only created high network traffic, but also left players with slower network connection with stuttering and jumping bars. This was solved by allowing the stat update code to be run on both sides. This creates a sort-of relationship between the client and server side. Since the client runs the same procedures, we can assume the values are in sync. Unfortunately, the client ticks are not always in sync with the server ticks. This may create a gap that, when considering butterfly effect, can kill you without you even knowing what's wrong. This is fixed by a synchronizing packet sent every 10 seconds. This reduces packets from 20 per second to 0.1 per second, or 20000% reduce in number of packets. Isn't this exhilarating?