Fixed strange sanity drain
Sanity was occasionally draining, but only on client. This was easily observable and reproducible bug, as it happened quite often. The cause was the client reported wrong values of light level, causing Survival Inc.'s sanity processor to thing the player is being in complete darkness (light level 0). This was fixed using a little inspiration from GuiOverlayDebug.