Added a meritocratic CoC
I value meritocracy, and I think it should also rule this
project. As such, I have decided to adopt the Code of Merit
as the code of conduct for this project. The Code of Merit
basically ensures that the project development doesn't steer
away from it's purpose, and that people's contributions are
determined by their technical value, not by the contributor's
individual attributes (described in the document).