Changed some season parameters into getters
Season parameters such as length and peak temperature are set in the config. As such, the enum parameters were replaced by getters that get the data from the mod config. Not only this removes the need to restart the game to change the values, but also it allows for pre-init usage of Season enum constants.