Phased out OverflowHandler
The only used overflow handler instance so far was the CAP instance. To remove unnecessary overhead, the overflow handlers were removed, and the checking is now performed by each individual StatProvider. These providers are themselves responsible to control their value outputs, as these are stored as-is in their respective StatRecords.