Improved client-side syncing capabilities
Until now, the StatSyncMessage was only evaluated for the local client player. But the problem arose when the client had to react to changes on other player's stats. This was introduced in this update. The StatSyncMessage now carries several trackers in it's payload, so that each individual tracker is sent to player, reflecting the server changes to the client world. For example, the stat command's effects on wetness are visible for other players as well.