Updated the ghost mechanics to the stat framework
This is a bit out-of-schedule commit. The ghost overhaul update is still in the clouds, but I had to do this. This commit basically phases out the old ghost capability, and replacing it with custom stat-based handler. This way, the synchronizations are done very easily, and thus making the effect crispier.