Moved stat provider instances back to their classes
The stat provider instances proved to be impractical when located in the common proxy class. As such, each stat provider got their own method called "init", in which they can initialize configuration-dependent stuff. As such, only the most basic stuff is done in constructors, while the more complex stuff is being done in the init method.