Added generic arguments to StatProvider
StatProviders now feature additional method called "getRecordClass",
which is used to cast the StatRecord and perform checks on the type.
Adding these generic arguments allows for better management of
StatProvider-requiring methods. That means that the traditional means of
checking the record type by creating a dummy record are no longer
required, nor supported.