Added ElementRenderFilter
ElementRenderFilters are general purpose filters designed to manipulate the underlying GUI render mechanism. It was designed for the purpose of moving the armor and air bubble bars up so that the stat bars don't just float when there is nothing below them. This makes the GUI a lot more visually appealing.