Logarithmic wetness drain function
Wetness is now drained in logarithmic manner. That means that the most significant losses in wetness are when it's value is high. Also, the slowdown was adjusted to use exponential-type function. This makes wetness effects way less annoying. Also, the brief slowdown pulses while swimming are no longer a thing.