Fixed #25
Fixed crash with AppleCore and HungerOverhaul. The crash was caused by AppleCore calling ItemCanteen.getItemUseAction() on dummy ItemStack instance with uninitialized NBT. This caused NPE when the code was attempting to get the canteen mode from the NBT. Confirmed that the crash no longer happens with AppleCore 3.4.0 and HungerOverhaul 1.3.3-jenkins148.