Improved initial stat synchronization
The initial stat synchronization was a process that often failed, leaving the user waiting up to 30 seconds for a synchronization event. This has been caused by the client player join event happening AFTER the server event. This has been addressed by adding another message-handler pair for synchronization requests. In other words, the client upon joining the world sends a message to the server requesting up-to-date data. The server replies with StatSyncMessage which contains data about all the players already online.