Simplified SimpleStatRecord#checkValue
The value checks done by checkValue method were simplified so that the
midpoint value doesn't need to be calculated. Moreover, the current
implementation is protected against IllegalStateException-s since the
code also checks whether the range has the endpoint in question
specified or not (i.e. whether the range is closed from that side)