Fixed sanity bar not rendered bug
A bug reported by curseforge user pigporker; The sanity bar was not rendered when Heat, Seasons and Ghost were disabled. Ultimately, this was caused by ill check in RenderHUD#isEffective (isUseful previously), which checked whether there is at least 1 element AND 1 filter. The sanity HUD upon itself doens't register any global filters, so the method was returning false and so the HUD was not registered as event handler.