Fixed crash when enabling stat module in-game
This was caused by the check being whether the module is *supposed* to be enabled or not, instead of whether it's actually loaded. This was fixed by introducing a static method to every provider checking whether it was initialized or not. Since many of the initialization calls were moved into the constructor, this may be causing some bugs, but I haven't spotted any yet.