Fixed #43
The problem was in the hastily implemented water quality module. Before, the server-side receiver didn't explicitly require RT details, since it only added constant value to the record. Now, with the water quality, RT results are always needed. But to overcome the need to do RT on server side (some mods enforce ridiculous policies on raytracing; see #34), the data is now being transferred in the packets themselves. As such, the independent server-side raytrace is completely optional, and thus is up to the user.