Reworked ItemSituationContainer event handling
This commit introduces some serious changes in order to address performance concerns and issue #46. The default catch-all event handler that was being used was terribly inefficient, since on every event (literally EVERY event, including rendering events that run multiple times per second) it had to run through all the situations and validate them. This was terribly inefficient, but the easiest to implement, since forge doesn't allow dynamic event listeners. This was addressed in this commit by introducing an ASM-Based system that generates event handler classes on demand. These dynamically generated event handlers then call the trigger method on the ItemSituationContainer class that created them. These classes need to be accessible to ASMEventHandler's class loader, and thus ForgeASMInjector was created for the exact purpose of defining classes inside this ClassLoader.