Added armor conductivity compatibility configuration
The armor conductivity mechanic was for a long time mostly limited to
vanilla armor materials due to the way how modded armor is registered.
In short, it was nearly impossible for players to discover the internal
enum name of the armor material. As such, they had a close to zero
chance of utilizing the primitive conductivity mapping in the main
config. In this commit, a whole new system was implemented based on
external JSON. In addition, players can now find the armor material ID
using the "/armorconductivity" command, which is able to both provide
the information regarding the already registered types and information
on the currently worn armor. Additionally, this command is able to
directly set conductivity modifiers of various armor materials with
immediate effect. These changes can be conveniently saved and/or loaded
from disk (using the aforementioned JSON file) by using the save/reload
functions of the command. In other words, setting an armor conductivity
for armor from other mods can now be performed entirely in-game using
few simple commands.