refactor: Removed static structures from stat providers
For a long time, stat providers were reliant on static methods such as init and loaded, together with the instance field. This effectively restricted them to singleton only usage, with not way of extending them if desired. This was resolved in this commit. Vast majority of the providers methods were changed to instance methods, and the provider instances were moved to SurvivalInc mod object. Thus, the instances can still be freely accessed, all while improving flexibility.