fix: Fixed duplicate canteen behavior
The canteen was giving duplicate penalties due to game engine bug where when item is updated, it is ticked again and thus the onItemUseFinish runs again. This commit fixed it by introducing a cooldown equal to the max use duration. Thus, it won't impact normal operation, but will filter out the duplicate operations.